First Commit
This script triggers on a Cloudwatch Autoscaling Event.  When that occurs, the script runs, checks to see if our main website's Elastic Beanstalk environment has any new instances.  If it does and these are not in our main Target Group, it adds that instance into it to begin serving traffic to the public.